requirements.txt
ollama
requests
supabase
flask
python-dotenv
beautifulsoup4
lxml
pydoll-python
Pillow
